Resolve "Backend - Student Search for Job Postings"

Closes #69

See merge request cs309/2024fall/mk_1_8!87

+++ Backend/TheHiveApp/src/main/java/com/example/thehiveapp/dto/jobPosting/JobPostingDto.java
@@ -20,7 +20,7 @@ public class JobPostingDto {
     private String summary;
     private BigDecimal salary;
     private JobType jobType;
-    private BigDecimal minimumGpa;
+    private Double minimumGpa;
     private LocalDate jobStart;
     private LocalDate applicationStart;
     private LocalDate applicationEnd;
